+++ Bootstrap.php
@@ -498,7 +498,7 @@ class Shopware_Plugins_Frontend_jtlconnector_Bootstrap extends Shopware_Componen
             (
               SELECT null, id, '', ordertime, '', invoice_amount, transactionID
               FROM s_order
-              WHERE LENGTH(transactionID) > 0
+              WHERE LENGTH(transactionID) > 0 AND cleared = 12
             )"
         );
     }
@@ -816,13 +816,14 @@ class Shopware_Plugins_Frontend_jtlconnector_Bootstrap extends Shopware_Componen
             CREATE TRIGGER `jtl_connector_payment` AFTER UPDATE ON `s_order`
             FOR EACH ROW
             BEGIN
-            IF LENGTH(NEW.transactionID) > 0 THEN
+            IF LENGTH(NEW.transactionID) > 0 AND NEW.cleared = 12 THEN
                     SET @paymentId = (SELECT id FROM jtl_connector_payment WHERE customerOrderId = NEW.id);
                     DELETE FROM jtl_connector_payment WHERE customerOrderId = NEW.id;
                     INSERT IGNORE INTO jtl_connector_payment VALUES (if(@paymentId > 0, @paymentId, null), NEW.id, '', now(), '', NEW.invoice_amount, NEW.transactionID);
                 END IF;
             END;
         ";
+
         Shopware()->Db()->query($sql);
     }
 }